multithreading and refactoring

+++ Test.Common/Models/CalculateNextNumberRequestModel.cs
@@ -5,7 +5,7 @@ namespace Test.Common.Models
     public class CalculateNextNumberRequestModel
     {
         [Required]
-        [Range(1, int.MaxValue)]
+        [Range(1, long.MaxValue)]
         public long CurrentNumber { get; set; }
     }
 }
